# inventory-api/docs/openapi/runtime-baseline.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "inventory-api runtime baseline",
    "version": "0.5.0",
    "description": "Baseline OpenAPI parcial y factual para superficies críticas verificadas en runtime. No cubre todo el monolito y se complementa con un manifiesto explícito de exclusiones contractuales."
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/clients/company": {
      "get": {
        "operationId": "listCompanyClients",
        "summary": "Listar clientes visibles de la empresa autenticada",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/company"
        },
        "responses": {
          "200": {
            "description": "Listado de clientes"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "operationId": "createCompanyClient",
        "summary": "Registrar cliente para la empresa autenticada",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/company"
        },
        "responses": {
          "201": {
            "description": "Cliente creado"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
    "/api/clients/classifications/company": {
      "get": {
        "operationId": "listCompanyClientClassifications",
        "summary": "Listar clasificaciones de clientes de la empresa autenticada",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/classifications/company"
        },
        "responses": {
          "200": {
            "description": "Listado de clasificaciones"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
    "/api/clients/document-types": {
      "get": {
        "operationId": "listClientDocumentTypes",
        "summary": "Listar tipos de documento soportados para clientes",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/document-types"
        },
        "responses": {
          "200": {
            "description": "Listado de tipos de documento"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
    "/api/clients/company/{clientId}/stores": {
      "post": {
        "operationId": "createCompanyClientStore",
        "summary": "Registrar tienda para un cliente visible de la empresa autenticada",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/company/:clientId/stores"
        },
        "responses": {
          "201": {
            "description": "Tienda creada"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
    "/api/clients/{clientId}/references": {
      "post": {
        "operationId": "createClientReference",
        "summary": "Registrar referencia de cliente",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/:clientId/references"
        },
        "responses": {
          "201": {
            "description": "Referencia creada"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
    "/api/clients": {
      "post": {
        "operationId": "createClientLegacyAlias",
        "summary": "Alias legacy preservado para crear clientes de compañía",
        "description": "Ruta histórica preservada por compatibilidad. Se prefiere `/api/clients/company` para documentación y consumidores nuevos.",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-compatibility": {
          "status": "legacy-alias",
          "preferredPath": "/api/clients/company",
          "removalStrategy": "preserve-until-broader-contract-coverage"
        },
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/"
        },
        "responses": {
          "201": {
            "description": "Cliente creado vía alias legacy"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
    "/api/clients/{id}": {
      "get": {
        "operationId": "getClient",
        "summary": "Consultar detalle de cliente",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/:id"
        },
        "responses": {
          "200": {
            "description": "Detalle de cliente"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      },
      "put": {
        "operationId": "updateClient",
        "summary": "Actualizar cliente visible para la empresa autenticada",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/:id"
        },
        "responses": {
          "200": {
            "description": "Cliente actualizado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      },
      "delete": {
        "operationId": "deleteClient",
        "summary": "Eliminar cliente desde rol admin de compañía",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/:id"
        },
        "responses": {
          "204": {
            "description": "Cliente eliminado"
          },
          "403": {
            "description": "Sin permisos admin"
          }
        }
      }
    },
    "/api/clients/{clientId}/documents": {
      "post": {
        "operationId": "uploadClientDocument",
        "summary": "Registrar documento privado de cliente",
        "description": "Superficie de payload alto justificado por carga base64 documentada.",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/:clientId/documents"
        },
        "responses": {
          "201": {
            "description": "Documento registrado"
          },
          "403": {
            "description": "Sin permisos"
          },
          "413": {
            "description": "Payload demasiado grande"
          }
        }
      }
    },
    "/api/clients/{clientId}/documents/{documentId}/download": {
      "get": {
        "operationId": "downloadClientDocument",
        "summary": "Descargar documento privado de cliente",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/clients",
          "routeFile": "src/routes/client.routes.js",
          "expressPath": "/:clientId/documents/:documentId/download"
        },
        "responses": {
          "200": {
            "description": "Descarga autorizada"
          },
          "403": {
            "description": "Sin permisos o fuera del tenant"
          },
          "404": {
            "description": "Documento no encontrado"
          }
        }
      }
    },
...
  }
}